Disable fullscreen to allow shaders to receive correct unifrom
SDL_FULLSCREEN flag must not be set so that shaders receive
correct uniforms. The texture_size sent to shaders in SDL1.2 gl
is always equal to the screen size if SDL_FULLSCREEN is set
when calling SDL_SetVideoMode. By not setting SDL_FULLSCREEN,
the correct texture_size is sent instead. SDL_SetVideoMode calls
PSP2_SetVideoMode, which enforces the SDL_FULLSCREEN flag, but
that happens after the correct texture_size is recorded.